Reorganize execution_policy.hpp underneath execution directory
