+++ agency/detail/bulk_functions/bulk_async_execution_policy.hpp
@@ -8,9 +8,9 @@
 #include <agency/detail/bulk_functions/bulk_invoke_execution_policy.hpp>
 #include <agency/detail/bulk_functions/shared_parameter.hpp>
 #include <agency/detail/bulk_functions/agent_shared_parameter_factory_tuple.hpp>
-#include <agency/detail/execution_policy_traits.hpp>
 #include <agency/execution/execution_agent.hpp>
 #include <agency/execution/executor/executor_traits.hpp>
+#include <agency/execution/detail/execution_policy_traits.hpp>
 #include <utility>
 
 namespace agency
