+++ agency/execution.hpp
@@ -4,4 +4,5 @@
 #include <agency/execution/execution_categories.hpp>
 #include <agency/execution/executor.hpp>
 #include <agency/execution/execution_agent.hpp>
+#include <agency/execution/execution_policy.hpp>
