+++ agency/execution/execution_policy.hpp
@@ -8,10 +8,10 @@
 #include <tuple>
 #include <initializer_list>
 
+#include <agency/detail/tuple.hpp>
 #include <agency/execution/executor.hpp>
 #include <agency/execution/execution_agent.hpp>
-#include <agency/detail/execution_policy_traits.hpp>
-#include <agency/detail/tuple.hpp>
+#include <agency/execution/detail/execution_policy_traits.hpp>
 
 namespace agency
 {
