+++ testing/bulk_invoke/multiple_results.cpp
@@ -1,5 +1,4 @@
-#include <agency/bulk_invoke.hpp>
-#include <agency/execution_policy.hpp>
+#include <agency/agency.hpp>
 #include <iostream>
 
 template<class ExecutionPolicy>
