+++ testing/bulk_then/multiple_results.cpp
@@ -1,5 +1,4 @@
-#include <agency/bulk_then.hpp>
-#include <agency/execution_policy.hpp>
+#include <agency/agency.hpp>
 #include <iostream>
 
 template<class ExecutionPolicy>
